seekweb_pos1/requirements.txt
PyQt5
mysql-connector-python
pywin32
pyserial
python-barcode
Pillow
pygame
qrcode
reportlab
pyinstaller
matplotlib==3.7.0
